requirements.txt
alembic==1.7.3
app==1.0.0
astroid==2.8.0
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
docopt==0.6.2
ecdsa==0.17.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
future==0.18.2
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
platformdirs==2.3.0
pluggy==0.13.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pyasn1==0.4.8
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.11.1
pyparsing==2.4.7
pytest==6.2.2
pytest-watch==4.2.0
python-dotenv==0.19.0
python-jose==3.3.0
python-jose-cryptodome==1.3.2
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.3.24
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
watchdog==2.1.5
Werkzeug==2.0.1
wrapt==1.12.1
